# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Assinix",
        "version": "1.0.0",
        "description": "Assinix \u00e9 uma API REST profissional em PHP/Laravel para gerenciamento de assinaturas, vencimentos e gastos recorrentes.\n\nA API contempla autentica\u00e7\u00e3o com Laravel Sanctum, CRUD de categorias, formas de pagamento e assinaturas, c\u00e1lculo financeiro, dashboard, cota\u00e7\u00e3o de moedas, auditoria, logs estruturados e documenta\u00e7\u00e3o OpenAPI."
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Local Docker"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/categories/{category}": {
            "get": {
                "operationId": "categories.show",
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Categoria retornada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "category": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "required": [
                                                "category"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "categories.update",
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Categoria atualizada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "category": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "required": [
                                                "category"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "categories.destroy",
                "tags": [
                    "Category"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Categoria removida com sucesso."
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/payment-methods/{paymentMethod}": {
            "get": {
                "operationId": "payment-methods.show",
                "tags": [
                    "PaymentMethod"
                ],
                "parameters": [
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "required": true,
                        "description": "The payment method ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Forma de pagamento retornada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "payment_method": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "required": [
                                                "payment_method"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "payment-methods.update",
                "tags": [
                    "PaymentMethod"
                ],
                "parameters": [
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "required": true,
                        "description": "The payment method ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePaymentMethodRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Forma de pagamento atualizada com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "payment_method": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "required": [
                                                "payment_method"
                                            ]
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "payment-methods.destroy",
                "tags": [
                    "PaymentMethod"
                ],
                "parameters": [
                    {
                        "name": "paymentMethod",
                        "in": "path",
                        "required": true,
                        "description": "The payment method ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Forma de pagamento removida com sucesso."
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "errors",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "PaymentMethodType": {
                "type": "string",
                "enum": [
                    "credit_card",
                    "debit_card",
                    "pix",
                    "bank_slip",
                    "bank_transfer",
                    "cash",
                    "other"
                ],
                "title": "PaymentMethodType"
            },
...
            "UpdateCategoryRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 100
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 500
                    },
                    "color": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "pattern": "^#[0-9A-Fa-f]{6}$"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "title": "UpdateCategoryRequest"
            },
...
            "UpdatePaymentMethodRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 100
                    },
                    "type": {
                        "$ref": "#/components/schemas/PaymentMethodType"
                    },
                    "brand": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 50
                    },
                    "last_four": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "title": "UpdatePaymentMethodRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}